server/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.2
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
bangla==0.0.2
blinker==1.8.1
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
ecdsa==0.19.0
einops==0.8.0
email_validator==2.1.1
encodec==0.1.1
etelemetry==0.3.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.2
filelock==3.14.0
Flask==3.0.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
g2pkk==0.1.2
grpcio==1.63.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
inflect==7.2.1
isodate==0.6.1
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.42.0
looseversion==1.3.0
lxml==5.2.1
marisa-trie==1.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.22.0
openai==1.25.1
openai-whisper==20231117
opencv-python==4.9.0.80
orjson==3.10.3
packaging==24.0
pandas==1.5.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==10.3.0
platformdirs==4.2.1
pooch==1.8.1
preshed==3.0.9
protobuf==5.26.1
prov==2.0.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydot==2.0.0
Pygments==2.17.2
pymongo==4.7.1
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pynndescent==0.5.12
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
pytesseract==0.3.10
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.11.4
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.37.2
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
tqdm==4.66.4
trainer==0.0.36
traits==6.3.2
transformers==4.40.1
TTS==0.22.0
typeguard==4.2.1
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
Werkzeug==3.0.2
yarl==1.9.4
